Replace use of Boost Format with fmt
Using the version of fmt that's bundled with spdlog.

The placeholder syntax has changed, and the new syntax is compatible
with C++20's std::format.

Most of the usage was together with translatable strings: the PO files
have been updated to reflect the placeholder syntax change. All
translatable strings now use explicitly-indexed placeholders for
flexibility and consistency.