Optionally download masterlists during install
Download and install all the latest masterlists (and prelude) when
installing LOOT, so that there is already a valid masterlist present
when LOOT starts.

All the masterlists are downloaded to avoid having to do game detection
at install time, and it means that if a game is installed later, LOOT
will already have a masterlist for it (though it may be outdated).

This is optional so that the installer does not require internet access,
but is enabled by default.