Add support for GOG distribution of Enderal SE
Unfortunately it uses a different local appdata folder than the Steam
distribution, and the Steam and GOG distributions both write the same
Registry key, so you can't tell which to use until you've checked what's
in the install path.

As such, I've moved the hardcoded local appdata folder for Enderal SE so
it's set during game detection instead, and only if there isn't already
a path defined (to avoid clobbering any custom path the user may have
provided). The presence of "Enderal Launcher.exe" is used to tell if
the Skyrim SE install is actually an Enderal SE install, and the
presence of "Galaxy64.dll" is used to tell if it's a GOG distribution.