Add support for multiple instances of each game
This adds support for using detecting multiple copies of a supported
game.

The supported games' registry keys have been split into source-specific
keys (e.g. Steam, GOG) and are now fully hardcoded - the configured
registry keys are now ignored, and I'll remove them in a later commit.

The basic approach is to loop through the supported games, checking all
the same places that LOOT checked before, but recording all found
installs instead of stopping at the first one found.

The tricky bit is merging those detected installs with LOOT's existing
configured games. That's done by first looking for equivalent game
install paths, and then looking for settings with an empty install path
and the same game type and sometimes the same folder name (because game
type is ambiguous for Oblivion, Skyrim, Nehrim and Enderal). Comparing
folder names is not always accurate, but it's more likely to give false
negatives than false positives, and I expect both to be very rare as
they'd require users to add custom settings for those games.

If a match is found, the settings object's install path is updated if it
doesn't match the detected path, and the local path is updated if it's
empty.

The remaining unmatched installs are used to create new game settings
objects, which are distinguished as necessary using unique names with
<source> and <index> suffixes to distinguish between copies of a game
installed from multiple sources (e.g. Steam, GOG) and between multiple
copies from a single source.

Unlike the rest of LOOT, game detection treats Nehrim, Enderal and
Enderal SE as separate games, as they've got different settings. This
makes a few things slightly awkward: it's easy to map a GameId (as
used by detection) to a GameType (as used by libloot and the rest of
LOOT), but to go the other way also needs a valid install path to check
if it's the original game or the total conversion mod. It might be worth
using GameId in GameSettings too, and leaving GameType for interfacing
with libloot (which doesn't care about the differences between the
original games and their total conversions).