Add support for Microsoft Store Fallout 4 DLCs
The Microsoft Store installs Fallout 4 DLCs to directories outside of
the game install path, and that version of the game knows to look for
plugins to load in those Data directories as well as the game's own
data directory.

To support this:

- Resolve plugin filenames to paths by checking if the plugin exists in
  one of the external Data directories, and use that path if so,
  otherwise return the plugin filename in the main Data path.
- When scanning for installed plugins, check the external Data
  directories first.

Data paths are checked in the same order that the game checks them.

It's safe to use relative paths from the game install path to the
external DLC paths, because the game will fail to launch if the DLC
plugins are moved or renamed.

Redating plugins has been updated to resolve plugin paths, but that
cannot be tested because redating only has an effect for Skyrim and
Skyrim SE, which have no external plugin paths.