Remove mutex in Game class
While the Game class does see some use across threads, the methods that
used the mutex do not get called from multiple threads concurrently, so
the mutex isn't really necessary. It's also only protecting a small
part of the data that's accessed in multiple threads, so gives false
confidence that the class is thread-safe.