Fix building on Linux with OneTBB 2021.8.0
The TBB library is used on Linux to provide the implementation of the
standard library parallel execution algorithms, but oneTBB 2021.8.0
defines an emit() function that clashes with Qt's emit macro, which may
be defined when game.h is included.

As such, undefine the macro before including the standard library
headers, and redefine it afterwards if it was previously defined.