Add support for reading Steam config to find games
This should avoid having to run game launchers before LOOT for users
who install games through Steam, and also means that LOOT is able to
find Steam games on Linux.

The VDF/ACF file parsing is very hacky, it works on my Steam installs
but I'm going to swap it out for an existing library, as Valve's
documentation of the file format isn't great so going for something
with an existing test suite and that's hopefully already in use will
probably give better results than a more complete DIY solution.