Link against ICU 71.1 on Linux
ICU 71.1 is provided by the Flatpak KDE runtime, so does not need to be
included in LOOT's Flatpak package.

ICU 66.1 still needs to be included in the Flatpak package as libloot's
latest release links against it, but the next release will link against
71.1, so it'll be possible to remove 66.1 from
the Flatpak archive after updating to the next libloot release.