Copy existing masterlist when initialising game folder
If multiple installs of a game are detected, LOOT will append unique
suffixes to their LOOT folder names, so the installer's
bundled/downloaded masterlists aren't used in that case.

To resolve this, try to copy the masterlist and its metadata TOML file
from the appropriate default LOOT folder for the game ID being
initialised if the game's folder doesn't have a masterlist.