Refer to override plugins as overlay plugins
xEdit, Wrye Bash and MO2 have all started using "overlay" to refer to
the new plugin type introduced by Starfield, which is a less ambiguous
term, so switch to it for consistency.

This doesn't make any changes to names in the code though, because that
would involve breaking API changes and we'll probably have to rename
again once we know what Bethesda call the plugin type.