Disable sorting for Starfield
It turns out that in Starfield it's not possible to determine the
overlap between two plugins without checking the light (and possibly
overlay) flags of their masters, as that affects how individual FormIDs
are resolved. Unfortunately, this means it's impossible to resolve
FormIDs correctly if any of a plugin's masters have had those flags
changed since the plugin was written, so it's not possible to reliably
calculate overlap in general. More info about this on Discord here:
https://discord.com/channels/471930020454072348/512837488922984458/1161962113070284800

Hopefully Bethesda will change this behaviour, otherwise it should be
possible to look up masters as part of overlap detection, but the latter
would require significant API changes and would still need to assume no
masters' flags have changed, which isn't a reliable assumption in Skyrim
SE or Fallout 4.

Until this issue is resolved, disable sorting so that LOOT can't
calculate load orders incorrectly.