Build Flatpak from source
The Flatpak is built in a separate CI job because it takes about twice
as long as the Linux job.

There's a new script that generates manifests for the Python and Rust
dependencies, as the Flatpak builder needs to download them ahead of
time but can't just run pip or cargo itself. The script must be run
before the Flatpak is built, and when the Flathub listing is updated the
dependency manifests will need to be committed to its repository since
Flathub's CI can't/won't run the script itself.

This includes a few improvements to how the Flatpak is built:

- --install-deps-from is used to avoid repeating the dependency versions
  unnecessarily
- Built modules are cached, since some now take several minutes.
- --ccache is used to enable caching of compiled object files, so even
  partial build results can be reused if their sources haven't changed.
- The CI job now calls the helper script to reduce duplication.