Update Qt to 6.7.0
This involves also updating ICU to 73.2 on Linux, as Qt 6.7.0 links
against ICU 73. ICU 73 binaries are no longer available for Ubuntu
20.04 and its glibc is too old to use with the 22.04 binaries, so
update the runner to Ubuntu 22.04. That in turn required updating to GCC
12 so that it could link with the updated TBB library installed by
Ubuntu.

This might cause runtime incompatibilities for the Flatpak, as I'm not
sure what glibc the KDE 6.7 platform provides, but that possibility will
be addressed in an upcoming commit.