Write Starfield.ccc on initialisation
Starfield's official plugins are implicitly active and are removed from plugins.txt when a saved game is loaded, so the only way to reliably set their load order is to set their file modification timestamps or to list them in Starfield.ccc.

Setting the file modification timestamps would mean they'd load after everything in plugins.txt unless hoisted to load earlier, which is difficult for LOOT to work with, and for Starfield.esm in particular loading it first is the obvious choice (if you change its load order slot, existing save games crash on load).

Putting the plugins in Starfield.ccc means that they load before everything in plugins.txt, which is more reliable over a range of load orders.

This replaces the existing content of Starfield.ccc because it's not created by default, and replacing any existing content avoids having to understand anything already in there. The only other utility known to write to Starfield.ccc is Mod Organizer 2, and the content LOOT writes is the same as what MO2 writes.