Remove Chinese translation
As it stands, the Chinese translation has 231 of 284 strings translated.

SilentDark is the only person to have contributed Chinese translations, and hasn't responded to attempts to get their permission to relicense the translations under the CC0 license. To avoid having a mix of translations under different licenses, I think it's best to remove their contributions (ideally temporarily).