Fix OpenMW detection on Linux
OpenMW is the only supported game that's available natively on Linux,
and unfortunately that means dealing with several variations on how it
can be packaged. In particular, when LOOT is running in a Flatpak
sandbox, it can't access the host's /etc and /usr paths, so it needs to
access their contents through /run/host/etc and /run/host/usr
respectively.