Run the Flatpak linter as part of the CI job
To reduce the chance of surprises when a release has been tagged and is
then submitted to Flathub.

The appstream-* errors are excepted as resolving them requires passing
an option during the build that only really makes sense when building
to be published to a registry, which the CI job doesn't do.

The other exceptions are duplicates of exceptions that are already
accepted by Flathub, but it looks like providing user exceptions causes
the linter to ignore all those already accepted by Flathub.