Use the libloot-rs reimplementation of libloot
It's got yet got any binary releases, so can only be built from source - fortunately that's a lot easier than with C++ libloot.

This builds libloot as part of the LOOT build in CI, but it might be worth going back to separately building and caching it, and maybe taking the same approach for Windows.