Use fusion as the default Qt style on Windows
Its advantages over the windowsvista style are:

- It supports following the system colour scheme (i.e. dark mode without
  having to select a different LOOT theme).
- It's unaffected by
  <https://qt-project.atlassian.net/browse/QTBUG-140898>, which is a
  regression in menu icon styling that has affected the windowsvista
  style since 6.9.2 and is still unfixed and present in 6.11.0.

The disadvantage is that it looks less native, but that's relatively
minor, particularly as looking native means very little on Windows,
which has GUI elements in something like 8 or 9 different styles.

The fusion style's advantages over the windows11 theme are:

- it's more internally consistent and more polished
- it can be used when running on Windows 10

On the other hand, it doesn't attempt to look native to Windows 11, but
then I'd argue that the windows11 style does a terrible job of that
anyway (particularly with a light colour scheme), and that it's better
to avoid it.

The fusion style is also the default for Linux desktops that use GNOME (KDE
uses the Breeze theme instead).