Work around windows11 and fusion (on Windows) dark colour scheme issues
This adds variations of the default-dark theme to work around two styling
issues present when using the windows11 or fusion Qt styles on Windows.
The two issues are:

1. When the system colour scheme is dark, the windows11 and fusion Qt styles
   use the system accent colour as the hyperlink text colour, which might not
   have enough contrast to be readable.
2. When the system colour scheme is dark, the fusion Qt style uses the system
   accent colour as the checkbox indicator's background colour in both checked
   and unchecked states, making it visually noisy and harder to distinguish.

   Unfortunately, overriding the background colour using QSS also removes the
   border styling, so the border needs to be restyled, but QSS doesn't support
   expressing the fusion theme's default styling, which derives a colour from
   the palette's dark colour:

   <https://github.com/qt/qtbase/blob/dev/src/widgets/styles/qfusionstyle.cpp#L587-L588>

   As a result the border styling is close to but not the same as default.

These issues don't occur when using the windowsvista Qt style because it
only has a light palette and ignores the system colour scheme. They also
don't occur on Linux (where the fusion style is available but the
windows11 style is not).