Tweak dark theme background and border colours
Using the colours of the default theme with the "windows" style and a dark system colour scheme as a guide, though there are some intentional differences. The result is darker overall but adds more contrast in some places while reducing contrast where it was excessive (e.g. the Bash Tag suggestions group box border).