Allow system colour scheme to be overridden by theme selection
If a theme has light and/or dark versions, the base theme name will be
offered alongside the light and/or dark versions in the dropdown, e.g.
the default theme appears as "default", "default (Dark)" and "default
(Light)".

Selecting a light or dark version of a theme will override the colour
scheme to match, while selecting the base theme name will select the
appropriate version for the current system colour scheme.