Autogenerate new game folder names
The LOOT Folder input in the new game dialog has been a source of confusion for some users for a long time.

The value is more of a LOOT implementation detail than anything most users should need to care about, so instead of prompting for it, generate it from the game ID, avoiding collisions with existing folders and config.

The game name isn't used because it might contain invalid characters or sequences, and I don't think it's worth dealing with the extra complexity that brings.