[WIP] Update libloot
WIP because the version used in CI and the Flatpak manifest is an unreleased version, and the release version in CMakeLists.txt hasn't yet been updated.

This updates Windows CI builds to build libloot from source so that it's easy to run CI builds against unreleased versions of libloot. As release builds are only done for Windows using MSVC and released versions of libloot, they continue to use the default prebuilt binaries.

The builds now download libloot by cloning its Git repository instead of using a source archive because using a Git repository allows the libloot build to embed its commit hash in the libloot binary.

This also modifies the Flatpak manifest to avoid line breaks in string literals because jq errors when it encounters them, it seems to want them to be escaped.