Support building libloot from source as part of the LOOT build
Update the CMake config to use first look for installed binaries, and if
not found then prebuilt binaries as the default libloot if building with
MSVC, otherwise use the Git repository if Git is installed, and
otherwise use a source archive. If the defaults are overridden, then a
non-default Git source is preferred over a non-default source archive
URL/path. There's no way to provide non-default prebuilt binaries.

Git repositories are preferred over source archives so that the libloot
build can embed the Git commit hash that it was built from in its binary
artifact.

Unfortunately due to name collisions when using Visual Studio,
FetchContent can't be used to handle building libloot from source, so a
more complex arrangement using ExternalProject and an imported library
target is used instead.