Add a "Show only new messages" filter
It hides all messages that existed the last time LOOT ran for the current game, before its content was last refreshed, or before its masterlist was last updated.

What counts as an old message may be updated even if switching to another game, updating the masterlist or refreshing content fails with an error.