Use uv lockfiles when generating flatpak manifests
The inline Python commands don't get a lockfile, but their pyyaml
dependency has no dependencies of its own, so a lockfile is unnecessary.

The Python version has been updated in CI because Python 3.11 includes a
toml library, and if that's not present then flatpak-pip-generator will
attempt to import a third-party library that's not declared as a
dependency.