Don't treat in-game downloads as official plugins
Although the removed plugins can be bought as part of a Creations Bundle DLC for Fallout 4 (on Steam and the Microsoft Store), they are only downloaded through the in-game Creations menu, alongside other Creations (there's also no way to just download the bundle's files, except to search for and download them one-by-one).

This doesn't really have any impact, as the official plugins list is only used for the cleaning message filter, it just makes it easier to manage this list (it only needs to change if a game update adds new plugins).